Refactor Header component to use async translations and remove LocaleSwitcherSelect. Update locale configuration and add localized labels for language selection in messages.
