+++ src/components/Header/Header.tsx
@@ -1,13 +1,11 @@
-"use client";
 import Image from "next/image";
 import React from "react";
 import { Button } from "../ui/button";
-import { useLocale, useTranslations } from "next-intl";
-import { LocaleSwitcherSelect } from "./LocaleSwitcherSelect";
+import { LocaleSwitcherSelect } from "../../features/LocaleSwitcher/LocaleSwitcherSelect";
+import { getTranslations } from "next-intl/server";
 
-const Header = () => {
-  const t = useTranslations("header");
-  const locale = useLocale();
+const Header = async () => {
+  const t = await getTranslations("header");
   return (
     <header className="max-w-6xl mx-auto px-6 py-6 flex items-center justify-between">
       <div className="flex items-center gap-3">
@@ -21,14 +19,7 @@ const Header = () => {
         />
       </div>
       <div className="flex items-center gap-3">
-        <LocaleSwitcherSelect
-          defaultValue={locale}
-          label="Language"
-          items={[
-            { value: "en", label: "English" },
-            { value: "uk", label: "Українська" },
-          ]}
-        />
+        <LocaleSwitcherSelect />
         <Button
           variant="outline"
           size="sm"
