+++ src/messages/uk.json
@@ -3,6 +3,9 @@
     "hero": {
       "title": "Аналізуй будь-який токен",
       "subtitle": "Отримуйте комплексні ризик-скоринги, зрозумілі описи та практичні інсайти для будь‑якого криптотокена."
+    },
+    "localeSwitcher": {
+      "label": "Мова"
     }
   },
   "analyzeForm": {
